initial commit: copy sources from SSD1289
I started project of new driver: SSD2119. This work is based on SSD1289.
I don't know exactly if this one is the most similar to my one. It's
just an experiment. There is no accompanying board file of any flavor,
but the hardware lying on my desk is Embest DM-STF4BB with DM-LCD35RT
LCD module.

Changes already made:
- Copied entire drivers/gdisp/SSD1289 directory into
  drivers/gdisp/SSD2119.
- Changed every occurrence of "SSD1289" to "SSD2119" in each source
  file.