Add GINPUT Dial, simplify GWIN input assignment
Added GINPUT Dial support and a driver that uses GADC to read the dial.
Added support for Dial inputs to the GWIN slider.
Updated the slider demo for Dial Inputs.
Simplified the assigning of inputs to GWIN "widgets" button and slider.
Updated the demo's to match the new input to button assignment.